requirements.txt
alembic==1.9.1
aniso8601==9.0.1
click==8.1.3
factory-boy==3.2.1
Faker==16.6.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
packaging==22.0
password-strength==0.0.3.post2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
testcase==0.1.0
Werkzeug==2.2.2
